backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-versatileimagefield==3.1
djangorestframework==3.15.2
pillow==11.0.0
python-magic==0.4.27
sqlparse==0.5.2
